backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
bcrypt>=4.1
requests>=2.31
numpy>=1.26,<2.0
scikit-learn>=1.4
pandas>=2.1.0

